+++ core/src/main/java/com/hujiugame/qingfeng/type/game/GameRenderPipeLineState.java
@@ -1,0 +1,7 @@
+package com.hujiugame.qingfeng.type.game;
+
+public enum GameRenderPipeLineState
+{
+    NORMAL,
+    FADING
+}